add MyTask and fix for m1 macbook

+++ src/lib/mixer_module/mixer_module.cpp
@@ -106,17 +106,17 @@ _control_latency_perf(perf_alloc(PC_ELAPSED, "control latency"))
 		_notch_filter_actuator[i+1].setParameters(_sample_freq, _param_imu_gyro_nf_freq.get(), _param_imu_gyro_nf_bw.get());
 	}
 	B_inv.setZero();
-	B_inv(0, 0)=-1.0f;
-	B_inv(0, 2)=1.0f;
+	B_inv(0, 0)=(double) -1.0f;
+	B_inv(0, 2)=(double) 1.0f;
 
-	B_inv(1, 1)=-1.0f;
-	B_inv(1, 2)=1.0f;
+	B_inv(1, 1)=(double) -1.0f;
+	B_inv(1, 2)=(double) 1.0f;
 
-	B_inv(2, 0)=1.0f;
-	B_inv(2, 2)=1.0f;
+	B_inv(2, 0)=(double) 1.0f;
+	B_inv(2, 2)=(double) 1.0f;
 
-	B_inv(3, 1)=1.0f;
-	B_inv(3, 2)=1.0f;
+	B_inv(3, 1)=(double) 1.0f;
+	B_inv(3, 2)=(double) 1.0f;
 
 	for (size_t i = 0; i < 4; i++)
 	{
@@ -511,17 +511,17 @@ bool MixingOutput::update()
 	{
 		if(_use_indi == 1)
 		{
-			_error_fb[0] = math::constrain(_controls[0].error_fb[actuator_controls_s::INDEX_ROLL], -1.f, 1.f);
-			_error_fb[1] = math::constrain(_controls[0].error_fb[actuator_controls_s::INDEX_PITCH], -1.f, 1.f);
-			_error_fb[2] = math::constrain(_controls[0].error_fb[actuator_controls_s::INDEX_YAW], -1.f, 1.f);
+			_error_fb[0] = (double) math::constrain(_controls[0].error_fb[actuator_controls_s::INDEX_ROLL], -1.f, 1.f);
+			_error_fb[1] = (double)math::constrain(_controls[0].error_fb[actuator_controls_s::INDEX_PITCH], -1.f, 1.f);
+			_error_fb[2] = (double)math::constrain(_controls[0].error_fb[actuator_controls_s::INDEX_YAW], -1.f, 1.f);
 
-			_indi_fb[0] = math::constrain(_controls[0].indi_fb[actuator_controls_s::INDEX_ROLL], -1.f, 1.f);
-			_indi_fb[1] = math::constrain(_controls[0].indi_fb[actuator_controls_s::INDEX_PITCH], -1.f, 1.f);
-			_indi_fb[2] = math::constrain(_controls[0].indi_fb[actuator_controls_s::INDEX_YAW], -1.f, 1.f);
+			_indi_fb[0] = (double)math::constrain(_controls[0].indi_fb[actuator_controls_s::INDEX_ROLL], -1.f, 1.f);
+			_indi_fb[1] = (double)math::constrain(_controls[0].indi_fb[actuator_controls_s::INDEX_PITCH], -1.f, 1.f);
+			_indi_fb[2] = (double)math::constrain(_controls[0].indi_fb[actuator_controls_s::INDEX_YAW], -1.f, 1.f);
 		}
-		_fb[0] = math::constrain(_controls[0].control[actuator_controls_s::INDEX_ROLL], -1.f, 1.f);
-		_fb[1] = math::constrain(_controls[0].control[actuator_controls_s::INDEX_PITCH], -1.f, 1.f);
-		_fb[2] = math::constrain(_controls[0].control[actuator_controls_s::INDEX_YAW], -1.f, 1.f);
+		_fb[0] = (double)math::constrain(_controls[0].control[actuator_controls_s::INDEX_ROLL], -1.f, 1.f);
+		_fb[1] = (double)math::constrain(_controls[0].control[actuator_controls_s::INDEX_PITCH], -1.f, 1.f);
+		_fb[2] = (double)math::constrain(_controls[0].control[actuator_controls_s::INDEX_YAW], -1.f, 1.f);
 
 		// float roll=0.0f;
 		// float pitch=0.0f;
@@ -554,7 +554,7 @@ bool MixingOutput::update()
 				dir_alloc_sim(_indi_fb, _uMin, _uMax, u_e, &z_e, &iters_e);
 				for (size_t i = 0; i < 3; i++)
 				{
-					double  temp = 0.0f;
+					double  temp = (double) 0.0f;
 					for(int k = 0 ; k < 4 ; k++)
 					{
 						temp += _B[i][k] * u_e[k];
@@ -576,7 +576,7 @@ bool MixingOutput::update()
 					dir_alloc_sim(_error_fb, uMin_new, uMax_new, u_d, &z_d, &iters_d);
 					for (size_t i = 0; i < 3; i++)
 					{
-						double  temp = 0.0f;
+						double  temp = (double) 0.0f;
 						for(int k = 0 ; k < 4 ; k++)
 						{
 							temp += _B[i][k] * u_d[k];
@@ -617,7 +617,7 @@ bool MixingOutput::update()
 
 		for (size_t i = 0; i < 3; i++)
 		{
-			double  temp = 0.0f;
+			double  temp = (double) 0.0f;
 			for(int k = 0 ; k < 4 ; k++)
 			{
 				temp += _B[i][k] * _u[k];
